Refactor: move AI providers under ai/, add travel module, add flight controller/service, remove .keep/.gitkeep files, update README
